requirements.txt
fastapi==0.111.0
pydantic==2.7.4
pandas==2.2.2
uvicorn==0.30.1
python-dotenv==1.0.0
pymongo==4.7.1
